Remove send_login_email from save method.
The change from commit 07faaeec causes two emails to be sent if we send
the email from the save method. This commit removes that from save and
makes it a separate explicit step. The only alternative would be to find
a way to pass our next url into the authenticate function but that seems
like it's getting messy.