chore: Refactor NoPasswordBackend
An authentication backend should return a user or None.
Returning a LoginCode does not comply with that spec.

Also, for some authentication backends it is a problem if no password is
given (e.g. allauth backend).
Checking if a user exists and creating a login code in the form should
be more stable this way.